backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
certifi==2025.10.5
django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ollama==0.6.0
pillow==12.0.0
pydantic==2.12.3
pydantic-core==2.41.4
sniffio==1.3.1
sqlparse==0.5.3
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
